ALL OF THE GATING

+++ resources/views/inputs/full.blade.php
@@ -1271,14 +1271,7 @@
 				<label for="collapsible_users" class="lbl-toggle card-header lbl-text"><i class="mr-2 fas fa-user"></i> User Management <a target="_blank" data-pt-position="right" data-pt-title='Fantasy Calendar Wiki: Settings' href='https://wiki.fantasy-calendar.com/index.php?title=User_Management' class="wiki protip"><i class="icon-question-sign"></i></a></label>
 				<div class="collapsible-content card-body">
 
-					@unless(Auth::user()->can('add-users', $calendar))
-
-						<div class='row no-gutters my-1'>
-							<p class='m-0'>Invite your friends to collaborate on this calendar!</p>
-							<p class='m-0'><a href="{{ route('subscription.pricing') }}" target="_blank">Subscribe now</a> to unlock this feature!</p>
-						</div>
-
-					@else
+					@if(Auth::user()->can('add-users', $calendar))
 
 						<div class='row no-gutters mt-1 mb-3'>
 							<p class='m-0'>Invite your friends to collaborate on this calendar! Once they accept your invite, you'll be able to assign them a role.</p>
@@ -1301,14 +1294,22 @@
 						<div class='row no-gutters my-1'>
 							<button type='button' class='btn btn-sm btn-secondary full' id='refresh_calendar_users'>Refresh</button>
 						</div>
-					@endunless
+
+					@else
+
+						<div class='row no-gutters my-1'>
+							<p>Invite your friends to collaborate on this calendar!</p>
+							<p class='m-0'><a href="{{ route('subscription.pricing') }}" target="_blank">Subscribe now</a> to unlock this feature!</p>
+						</div>
+
+					@endif
 
 				</div>
 			</div>
 
 		@endif
 
-		@if(request()->is('calendars/*/edit') && Auth::user()->can('link', $calendar))
+		@if(request()->is('calendars/*/edit'))
 			<!---------------------------------------------->
 			<!------------------ LINKING ------------------->
 			<!---------------------------------------------->
@@ -1317,28 +1318,40 @@
 				<label for="collapsible_linking" class="lbl-toggle card-header lbl-text"><i class="mr-2 fas fa-link"></i> Calendar Linking <a target="_blank" data-pt-position="right" data-pt-title='Fantasy Calendar Wiki: Calendar Linking' href='https://wiki.fantasy-calendar.com/index.php?title=Calendar_Linking' class="wiki protip"><i class="icon-question-sign"></i></a></label>
 				<div class="collapsible-content card-body">
 
-					<div id='calendar_link_hide'>
+					@if(Auth::user()->can('link', $calendar))
 
-						@if($calendar->parent != null)
-                            <div class='row no-gutters my-1 center-text hidden calendar_link_explaination'>
-                                <p class='m-0'>This calendar is already linked to a <a href='/calendars/{{ $calendar->parent->hash }}/edit' target="_blank">parent calendar</a>. Before linking any calendars to this one, you must unlink this calendar from its parent.</p>
-                            </div>
-                        @else
-                            <div class='row no-gutters my-1 center-text'>
-                                <p>Calendar linking is a complex feature - we recommend you check out the <a href='https://wiki.fantasy-calendar.com/index.php?title=Calendar_Linking' target="_blank"><i class="icon-question-sign"></i> Fantasy-Calendar wiki article</a> on the feature!</p>
-                            </div>
+						<div id='calendar_link_hide'>
 
-                            <div class='row no-gutters my-1'>
-                                <select class='form-control' id='calendar_link_select'></select>
-                            </div>
-                            <div class='row no-gutters my-1'>
-                                <button type='button' class='btn btn-sm btn-secondary full' id='refresh_calendar_list_select'>Refresh</button>
-                            </div>
+							<div class='row no-gutters my-1'>
+								<p>Calendar linking is a complex feature, we recommend you check out the wiki article on <a href='https://wiki.fantasy-calendar.com/index.php?title=Calendar_Linking' target="_blank"><i class="icon-question-sign"></i> Calendar Linking</a>.</p>
+							</div>
 
-                            <div class='sortable' id='calendar_link_list'></div>
-                            <div class='sortable mt-1' id='calendar_new_link_list'></div>
-                        @endif
-                    </div>
+							@if($calendar->parent != null)
+								<div class='row no-gutters my-1 center-text hidden calendar_link_explaination'>
+									<p class='m-0'>This calendar is already linked to a <a href='/calendars/{{ $calendar->parent->hash }}/edit' target="_blank">parent calendar</a>. Before linking any calendars to this one, you must unlink this calendar from its parent.</p>
+								</div>
+							@else
+
+								<div class='row no-gutters my-1'>
+									<select class='form-control' id='calendar_link_select'></select>
+								</div>
+								<div class='row no-gutters my-1'>
+									<button type='button' class='btn btn-sm btn-secondary full' id='refresh_calendar_list_select'>Refresh</button>
+								</div>
+
+								<div class='sortable' id='calendar_link_list'></div>
+								<div class='sortable mt-1' id='calendar_new_link_list'></div>
+							@endif
+						</div>
+
+					@else
+
+						<div class='row no-gutters my-1'>
+							<p>Link calendars together, and make this calendar's date drive the date of other calendars!</p>
+							<p class='m-0'><a href="{{ route('subscription.pricing') }}" target="_blank">Subscribe now</a> to unlock this feature!</p>
+						</div>
+
+					@endif
 				</div>
 			</div>
 		@endif
